[Dtype] Add FP8_E4M3 for DeepSeek. (#79)
* [Dtype] Add FP8_E4M3 for DeepSeek.

* [DeepSeek] Change tokenizer from fp16 to bf16.

+++ src/models/deepseek.h
@@ -49,7 +49,7 @@ private:
     void setFinalLnWeight(const std::string &modelPath);
 
 private:
-    TokenEmbedding<float16_t> *embedding;
+    TokenEmbedding<bfloat16_t> *embedding;
     RmsNorm finalLN;
 };
